+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/listener/PolarisConfigRefreshOptimizationListener.java
@@ -28,6 +28,7 @@ import com.tencent.cloud.polaris.config.enums.RefreshType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import org.springframework.beans.BeansException;
 import org.springframework.beans.factory.config.ConstructorArgumentValues;
 import org.springframework.beans.factory.support.AbstractBeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionBuilder;
@@ -65,17 +66,24 @@ public class PolarisConfigRefreshOptimizationListener implements ApplicationList
 
 	private static final String REFRESH_CONTEXT_REFRESHER_BEAN_NAME = "polarisRefreshContextPropertySourceAutoRefresher";
 
+
 	@Override
 	public void onApplicationEvent(@NonNull ContextRefreshedEvent event) {
 		ConfigurableApplicationContext applicationContext = (ConfigurableApplicationContext) event.getApplicationContext();
-		PolarisConfigRefreshScopeAnnotationDetector detector = applicationContext
-				.getBean(PolarisConfigRefreshScopeAnnotationDetector.class);
+		PolarisConfigRefreshScopeAnnotationDetector detector = applicationContext.getBean(PolarisConfigRefreshScopeAnnotationDetector.class);
 		boolean isRefreshScopeAnnotationUsed = detector.isRefreshScopeAnnotationUsed();
 		String annotatedRefreshScopeBeanName = detector.getAnnotatedRefreshScopeBeanName();
+		// using System.setProperty to set spring.cloud.polaris.config.refresh-type
+		String value = System.getProperty("spring.cloud.polaris.config.refresh-type");
+		boolean isSystemSetRefreshType = RefreshType.REFRESH_CONTEXT.toString().equalsIgnoreCase(value);
 		// a bean is using @RefreshScope, but the config refresh type is still [reflect], switch automatically
-		if (isRefreshScopeAnnotationUsed) {
-			LOGGER.warn("Detected that the bean [{}] is using @RefreshScope annotation, but the config refresh type is still [reflect]. "
-					+ "[SCT] will automatically switch to [refresh_context].", annotatedRefreshScopeBeanName);
+		if (isRefreshScopeAnnotationUsed || isSystemSetRefreshType) {
+			if (isRefreshScopeAnnotationUsed) {
+				LOGGER.warn("Detected that the bean [{}] is using @RefreshScope annotation, but the config refresh type is still [reflect]. " + "[SCT] will automatically switch to [refresh_context].", annotatedRefreshScopeBeanName);
+			}
+			if (isSystemSetRefreshType) {
+				LOGGER.warn("Detected that using System.setProperty to set spring.cloud.polaris.config.refresh-type = refresh_context, but the config refresh type is still [reflect]. " + "[SCT] will automatically switch to [refresh_context].");
+			}
 			switchConfigRefreshTypeProperty(applicationContext);
 			modifyPolarisConfigPropertiesBean(applicationContext);
 			// remove related bean of type [reflect]
@@ -89,8 +97,7 @@ public class PolarisConfigRefreshOptimizationListener implements ApplicationList
 
 	private void switchConfigRefreshTypeProperty(ConfigurableApplicationContext applicationContext) {
 		MutablePropertySources propertySources = applicationContext.getEnvironment().getPropertySources();
-		propertySources.addFirst(new MapPropertySource(CONFIG_REFRESH_TYPE_PROPERTY,
-				Collections.singletonMap(POLARIS_CONFIG_REFRESH_TYPE, RefreshType.REFRESH_CONTEXT)));
+		propertySources.addFirst(new MapPropertySource(CONFIG_REFRESH_TYPE_PROPERTY, Collections.singletonMap(POLARIS_CONFIG_REFRESH_TYPE, RefreshType.REFRESH_CONTEXT)));
 	}
 
 	private void modifyPolarisConfigPropertiesBean(ConfigurableApplicationContext applicationContext) {
@@ -99,9 +106,15 @@ public class PolarisConfigRefreshOptimizationListener implements ApplicationList
 	}
 
 	private void removeRelatedBeansOfReflect(ConfigurableApplicationContext applicationContext) {
-		DefaultListableBeanFactory beanFactory = (DefaultListableBeanFactory) applicationContext.getBeanFactory();
-		beanFactory.removeBeanDefinition(REFLECT_REFRESHER_BEAN_NAME);
-		beanFactory.removeBeanDefinition(REFLECT_REBINDER_BEAN_NAME);
+		try {
+			DefaultListableBeanFactory beanFactory = (DefaultListableBeanFactory) applicationContext.getBeanFactory();
+			beanFactory.removeBeanDefinition(REFLECT_REFRESHER_BEAN_NAME);
+			beanFactory.removeBeanDefinition(REFLECT_REBINDER_BEAN_NAME);
+		}
+		catch (BeansException e) {
+		    // If there is a removeBean exception in this code, do not affect the main process startup. Some user usage may cause the polarisReflectPropertySourceAutoRefresher to not load, and the removeBeanDefinition will report an error
+			LOGGER.debug("removeRelatedBeansOfReflect occur error:", e);
+		}
 	}
 
 	private void registerRefresherBeanOfRefreshContext(ConfigurableApplicationContext applicationContext) {
@@ -118,9 +131,9 @@ public class PolarisConfigRefreshOptimizationListener implements ApplicationList
 		beanFactory.registerBeanDefinition(REFRESH_CONTEXT_REFRESHER_BEAN_NAME, beanDefinition);
 	}
 
+
 	private void addRefresherBeanAsListener(ConfigurableApplicationContext applicationContext) {
-		PolarisRefreshEntireContextRefresher refresher = (PolarisRefreshEntireContextRefresher) applicationContext
-				.getBean(REFRESH_CONTEXT_REFRESHER_BEAN_NAME);
+		PolarisRefreshEntireContextRefresher refresher = (PolarisRefreshEntireContextRefresher) applicationContext.getBean(REFRESH_CONTEXT_REFRESHER_BEAN_NAME);
 		applicationContext.addApplicationListener(refresher);
 	}
 }