+++ CHANGELOG.md
@@ -26,3 +26,4 @@
 - fix:fix feign report -1 port bug.
 - fix:upgrade spring version.
 - fix:Update README-zh.md
+- feature: support Polaris configuration center extension plugin interface and support dynamic modification of log levels.